requirements.txt
Flask
openai
rembg
Pillow
